+++ src/utils/i18n/locales/bod/bod.json
@@ -17,5 +17,13 @@
         "pre": "ནང་པའི་ཤེས་རབ་སྲུང་སྐྱོབ་ཀྱི་ཆེད་དུ་བྱམས་སེམས་ཀྱིས་བཟོས་པ།",
         "post": ""
     },
-    "historyTitle": "བཀའ་འགྱུར་གྱི་ལོ་རྒྱུས།"
+    "historyTitle": "བཀའ་འགྱུར་གྱི་ལོ་རྒྱུས།",
+    "earlyTranslationPeriod": "སྔ་དར་ལོ་ཙཱ་བའི་དུས་སྐབས།",
+    "darkAgeAndRevival": "བསྟན་པ་ཉམས་པ་དང་ཡང་སྐྱེས།",
+    "protoKangyurFormation": "བཀའ་འགྱུར་འདས་པའི་གྲུབ་པ།",
+    "firstStructuredKangyurs": "བཀའ་འགྱུར་གྱི་རིམ་པ་དང་པོ།",
+    "classicManuscriptKangyurs": "དཔེ་ཆའི་བཀའ་འགྱུར་ཆེན་མོ།",
+    "blockPrintedEditions": "པར་རྐྱང་བཀའ་འགྱུར།",
+    "twentiethCenturyEditions": "ཆེ་བར་དུས་རབས་ཀྱི་པར་མ།",
+    "twentyFirstCenturyDigitalEditions": "ཉེར་གཅིག་དུས་རབས་བརྙན་རིག་པར་མ།"
 }
