+++ src/utils/i18n/locales/en/en.json
@@ -17,5 +17,14 @@
         "pre": "Made with",
         "post": "for the preservation of Buddhist wisdom"
     },
-    "historyTitle": "History of the Kangyur"
+    "historyTitle": "History of the Kangyur",
+    "earlyTranslationPeriod": "Early Translation Period",
+    "darkAgeAndRevival": "Dark Age and Revival",
+    "protoKangyurFormation": "Proto-Kangyur Formation",
+    "firstStructuredKangyurs": "First Structured Kangyurs",
+    "classicManuscriptKangyurs": "Classic Manuscript Kangyurs",
+    "blockPrintedEditions": "Block-Printed Editions",
+    "twentiethCenturyEditions": "20th Century Editions",
+    "twentyFirstCenturyDigitalEditions": "21st Century Digital Editions"
+
 }
